+++ R/calc_route_mc.R
@@ -1,118 +1,150 @@
-#' Calc migratory connectivity from BirdFlowR routes
+# nolint start: object_name_linter
+#' Calculate migratory connectivity from BirdFlowR routes.
 #'
-#' **WARNING** this function is experimental and new and likely to change
-#' a bunch.
+#' This function calculates migratory connectivity based on sample of routes
+#' objects sampled either from a BirdFlowR model or from observational track
+#' data.
 #'
-#' Currently, it assumes that all the routes start and end at the same
-#' timestep and so uses the first and last timestep of the first route as the
-#' starting and ending timesteps and then calculates the migratory connectivity
-#' between those timesteps across all routes, dropping routes that don't
-#' have both timesteps.
+#' It calculates the migratory connectivity metric MC as defined in Cohen 2018.
+#' MC represents an abundance-weighted correlation that is calculated between
+#' the origin locations and target locations,
+#' taking into account all grid transitions for the specified time period.
 #'
-#' It treats the routes as "Telemetry" information when calling
-#' `MigConnectivity` functions.
+#' The route implementation does not correct for spatial sampling imbalances
+#' of the provided routes,
+#' i.e. the transition matrix is calculated directly from the route transitions.
 #'
-#' Current process is to call `MigConnectivity::estPsi` internally to calculate
-#'  the transition probabilities needed to calculate `MC`.
-#'  This step is really slow and then pass the result to `calcMC`.
-#'
-#'  It might make sense to instead call  `estMC` which I think would perform
-#'  both steps at once and output uncertainty.
-#'
-#'  Alternatively, we could see if we can call a more efficient function to
-#'  calculate the transitions probabilities as `estPsi` is doing a lot of
-#'  work to calculate uncertainties that aren't used with the current process.
-#'
-#'
+#' When sampling a high number of routes from a BirdFlow model, the output of
+#' [calc_route_mc()] will become asymptotically identical to the output of
+#' [calc_birdflow_mc()].
 #'
 #' @param rts Output from [BirdFlowR::route()]
-#' @param bf  The BirdFlow model used to make `rts`
+#' @param bf The BirdFlow model used to make `rts`
+#' @param exact logical. Whether to match route time steps exactly to period
+#' requested with [BirdFlowR::lookup_timestep_sequence()] (TRUE)
+#' or use the route's closest available time steps (FALSE).
+#' @param delta_steps The number of time steps (weeks) you allow for the nearest
+#' timestep search when exact = FALSE. Default value is 2.
+#' @inheritDotParams BirdFlowR::lookup_timestep_sequence -x
 #' @return migratory connectivity estimated from the `rts` object.
 #' @export
-calc_route_mc <- function(rts, bf) {
+#' @seealso [calc_birdflow_mc()]
+#' @examples
+#' bf <- BirdFlowModels::amewoo
+#' # generate 100 synthetic routes
+#' rts <- BirdFlowR::route(bf, 100, season = "prebreeding")
+#' # calculate MC across prebreeding season
+#' calc_route_mc(rts, bf, season="prebreeding")
+#' # calculate MC across a subset of weeks:
+#' calc_route_mc(rts, bf, start=10, end=20)
+#' # set exact to false to not enforce exact matches of route timestamps and
+#' # requested start and end weeks (in this example end week 30 is after the
+#' # last timestamp of the input routes):
+#' calc_route_mc(rts, bf, start=10, end=30, exact=FALSE)
+#' @references
+#' Cohen EB, Hostetler JA, Hallworth MT, Rushing CS, Sillett TS, Marra PP.
+#' Quantifying the strength of migratory connectivity.
+#' Methods in Ecology and Evolution. 2018 Mar;9(3):513-24.
+#' \doi{10.1111/2041-210X.12916}
+#' @importFrom rlang .data
+calc_route_mc <- function(rts, bf, exact = TRUE, delta_steps = 2, ...) {
   # Using MigConnectivity::estPSI treating the routes as tracking data
+  stopifnot(is.logical(exact))
 
-
-  ts <- rts$timestep[rts$route_id == rts$route_id[1]]
+  ts <- lookup_timestep_sequence(bf, ...)
   origin_t <- ts[1]
   target_t <- ts[length(ts)]
-  ids <- unique(rts$route_id)
+  ids <- unique(rts$data$route_id)
 
-  origin <- rts[rts$timestep %in% origin_t, ]
-  target <- rts[rts$timestep %in% target_t,]
+  nearest_timestep <- function(rts, tstep, delta_steps) {
+    rts$data |>
+      dplyr::group_by(.data$route_id) |>
+      dplyr::mutate(dt = abs(.data$timestep - tstep)) |>
+      # must be within delta_steps of tstep:
+      dplyr::filter(.data$dt <= delta_steps) |>
+      dplyr::slice_min(.data$dt, with_ties = FALSE) |>
+      dplyr::select(-.data$dt) |>
+      dplyr::ungroup() |>
+      as.data.frame()
+  }
+
+  if (exact) {
+    origin <- rts$data[rts$data$timestep %in% origin_t, ]
+    target <- rts$data[rts$data$timestep %in% target_t, ]
+  } else {
+    origin <- nearest_timestep(rts, origin_t, delta_steps)
+    target <- nearest_timestep(rts, target_t, delta_steps)
+  }
 
   # Make sure origin and target rows correspond
   # Not necessary for synthetic routes because likely needed for real routes
   retained_ids <- intersect(origin$route_id, target$route_id)
-  origin <- origin[match(origin$route_id, retained_ids), ]  |> as.data.frame()
-  target <- target[match(target$route_id, retained_ids), ]  |> as.data.frame()
-  stopifnot(isTRUE(all.equal(origin$route_id, retained_ids)))
-  stopifnot(isTRUE(all.equal(target$route_id, retained_ids)))
-  make_pts <- function(x){
-    x$geom <- vector(mode = "list", length = nrow(x))
-    for(i in seq_len(nrow(x))){
-      x$geom[[i]] <- sf::st_point(as.numeric(x[i, c("x","y")]))
-    }
-    res <- sf::st_as_sf(x, sf_column_name = "geom")
+  dropped_ids <- ids[!ids %in% retained_ids]
+  if (length(dropped_ids) > 0) {
+    warning(paste("dropping", length(dropped_ids),
+                  "tracks not connecting start and end period ...",
+                  length(retained_ids), "tracks remaining"))
   }
-
-  origin_pts <- make_pts(origin)
-  target_pts <- make_pts(target)
-
-  sf::st_crs(origin_pts) <- sf::st_crs(bf)
-  sf::st_crs(target_pts) <- sf::st_crs(bf)
+  origin <- origin[origin$route_id %in% retained_ids, ] |> as.data.frame()
+  target <- target[target$route_id %in% retained_ids, ] |> as.data.frame()
+  stopifnot(isTRUE(setequal(origin$route_id, retained_ids)))
+  stopifnot(isTRUE(all.equal(target$route_id, retained_ids)))
 
 
-  # It seems that MigConnectivity needs origin and target assignment to be 1:n
-  # Here I make a crosswalk between i and
-  # a new id that ranges 1:n.  I also include a name that is "i" combined with
-  # the value of i for each cell.
-  # Origin
-  origin_sites <- data.frame(i = sort(unique(origin$i)))
-  origin_sites$name <- paste0("i", origin_sites$i)
-  origin_sites$id <- seq_len(nrow(origin_sites))
-  # Add new ID and name to origin
-  mv <- match(origin_pts$i, origin_sites$i)
-  origin_pts$id <- origin_sites$id[mv]
-  origin_pts$name <- origin_sites$name[mv]
-  # Target
-  target_sites <- data.frame(i = sort(unique(target$i)))
-  target_sites$name <- paste0("i", target_sites$i)
-  target_sites$id <- seq_len(nrow(target_sites))
-  # Add new ID and name to target
-  mv <- match(target_pts$i, target_sites$i)
-  target_pts$id <- target_sites$id[mv]
-  target_pts$name <- target_sites$name[mv]
+  # Initialize origin and target distributions
+  origin_abun <- target_abun <- rep(0, n_active(bf))
+  # Tally the origin and target cells
+  target_table <- table(target$i)
+  origin_table <- table(origin$i)
+  # Populate origin and target distributions
+  target_abun[as.numeric(names(target_table))] <-
+    target_table / sum(target_table)
+  origin_abun[as.numeric(names(origin_table))] <-
+    origin_table / sum(origin_table)
 
-  # Calculate PSI --- SLOW!
-  bf_msg("Calculating psi (transition probabilities) from routes.",
-         " This is slow!\n")
-  psi <- MigConnectivity::estPsi(originPoints = origin_pts,
-                                 targetPoints = target_pts,
-                                 originAssignment = origin_pts$id,
-                                 targetAssignment = target_pts$id,
-                                originNames = origin_sites$name,
-                                 targetNames = target_sites$name,
-                                 isTelemetry = rep(TRUE, nrow(origin_pts)))
+  # Calculate distance matrices
+  dist <- great_circle_distances(bf)  # all active cells
 
+  # matrix product to have all combinations for two samples for origin_abun
+  # note that this outer product remains normalized, because origin_abun is normalized
+  origin_abun_prod <- origin_abun %*% t(origin_abun)
+  # mu_D effectively equals weighted.mean(origin_dist, origin_abun_prod)
+  mu_D <- sum(dist * origin_abun_prod)
+  # SD in origin distance between any two given pixels
+  sd_D <- sqrt(sum((dist - mu_D)^2 * origin_abun_prod))
 
-  # Make distance matrices that correspond with origin_sites and target_sites
-  bf_dist <- great_circle_distances(bf)
-  origin_dist <- bf_dist[origin_sites$i, origin_sites$i]
-  target_dist <- bf_dist[target_sites$i, target_sites$i]
-  origin_abund <- get_distr(bf, origin_t)[origin_sites$i]
+  # same calculation, now for target_abun
+  target_abun_prod <- target_abun %*% t(target_abun)
+  mu_V <- sum(dist * target_abun_prod)
+  sd_V <- sqrt(sum((dist - mu_V)^2 * target_abun_prod))
 
-  # Re-standarize to sum to 1 - necessary but maybe problematic
-  origin_abund <- origin_abund / sum(origin_abund)
+  # construct transition matrix
+  psi <- as.matrix(table(origin$i, target$i))
+  # normalize the matrix
+  psi <- psi / rowSums(psi)
 
-  # Calculate migratory connectivity
-  bf_msg("Calculating migratory connectivity.\n")
-  mc <- MigConnectivity::calcMC(originDist = origin_dist,
-                                targetDist = target_dist,
-                                psi = psi$psi$mean,
-                                originRelAbund =  origin_abund)
+  origin_abun <- origin_abun[as.numeric(rownames(psi))]
+  target_abun <- target_abun[as.numeric(colnames(psi))]
 
+  # multiply transition matrix and relative abundance
+  psi_abun <- psi * origin_abun
 
-  return(mc)
+  # standardizing origin distance matrix
+  origin_std <-
+    (dist[as.numeric(rownames(psi)), as.numeric(rownames(psi))] - mu_D) / sd_D
+  target_std <-
+    (dist[as.numeric(colnames(psi)), as.numeric(colnames(psi))] - mu_V) / sd_V
+
+  # reduce the distance matrix and abundance vectors
+  dist <- dist[as.numeric(colnames(psi)), as.numeric(rownames(psi))]
+
+  dim(psi_abun)
+  dim(origin_std)
+  dim(target_std)
+
+  # calculate MC
+  MC <- sum(t(psi_abun) %*% origin_std %*% psi_abun * target_std)
+
+  return(MC)
 }
-
+# nolint end